# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UltraRank API - ultrarunner rankings and race results
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /runners:
    get:
      operationId: listRunners
      tags: [runners]
      summary: List all runners with optional filters
      parameters:
        - name: gender
          in: query
          schema:
            type: ["string", "null"]
        - name: country
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of runners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Runner"
    post:
      operationId: createRunner
      tags: [runners]
      summary: Create a new runner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunnerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Runner"
  /runners/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [runners]
      summary: Global leaderboard sorted by rating
      parameters:
        - name: gender
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Ranked list of runners
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
  /runners/{id}:
    get:
      operationId: getRunner
      tags: [runners]
      summary: Get a runner by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Runner profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Runner"
        "404":
          description: Not found
    patch:
      operationId: updateRunner
      tags: [runners]
      summary: Update a runner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunnerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Runner"
    delete:
      operationId: deleteRunner
      tags: [runners]
      summary: Delete a runner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /runners/{id}/results:
    get:
      operationId: getRunnerResults
      tags: [runners]
      summary: Get all race results for a runner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Runner's results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResultWithRace"
  /runners/{id}/stats:
    get:
      operationId: getRunnerStats
      tags: [runners]
      summary: Get aggregate stats for a runner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Runner stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunnerStats"
...
  /portal/scrape:
    post:
      operationId: portalScrapePreview
      tags: [portal]
      summary: Scrape and preview results from a URL without saving
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrapeInput"
      responses:
        "200":
          description: Scraped results preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapePreview"
        "422":
          description: Scrape failed or site not supported
  /portal/scrape-import:
    post:
      operationId: portalAutoCreateRaceImport
      tags: [portal]
      summary: Create a race and import results from a URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  type: string
                  format: uri
      responses:
        "200":
          description: Import successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  resultsCreated:
                    type: integer
                  runnersCreated:
                    type: integer
                  runnersUpdated:
                    type: integer
                  raceId:
                    type: integer
                  raceName:
                    type: string
                  source:
                    type: string
  /portal/races/{id}/scrape-import-async:
    post:
      operationId: portalScrapeImportAsync
      tags: [portal]
      summary: Start a background import job — returns a job ID immediately
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrapeInput"
      responses:
        "202":
          description: Job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsyncJobStarted"
        "422":
          description: Scrape failed
...
  /portal/jobs/{jobId}:
    get:
      operationId: portalGetJob
      tags: [portal]
      summary: Poll the status of a background import job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJob"
        "404":
          description: Job not found or expired
  /portal/races/{id}/auto-search:
    post:
      operationId: portalAutoSearch
      tags: [portal]
      summary: Search the web for results pages matching a race's name and date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of candidate result pages found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoSearchResponse"
        "403":
          description: Not your race
        "404":
          description: Race not found
  /portal/races/{id}/auto-import:
    post:
      operationId: portalAutoImport
      tags: [portal]
      summary: Auto-search and import the best-confidence results for a race
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Import successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeImportResponse"
        "403":
          description: Not your race
        "404":
          description: Race not found
        "422":
          description: No confident results found
components:
  schemas:
...
    Runner:
      type: object
      required: [id, name, country, gender, rating, rank, totalRaces]
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
        countryCode:
          type: ["string", "null"]
        gender:
          type: string
          enum: [M, F, X]
        age:
          type: ["integer", "null"]
        ageCategory:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        rating:
          type: number
        rank:
          type: integer
        totalRaces:
          type: integer
        totalDistanceKm:
          type: number
        bestFinish:
          type: ["integer", "null"]
        createdAt:
          type: string
    RunnerInput:
      type: object
      required: [name, country, gender]
      properties:
        name:
          type: string
          minLength: 1
        country:
          type: string
        countryCode:
          type: ["string", "null"]
        gender:
          type: string
          enum: [M, F, X]
        age:
          type: ["integer", "null"]
        bio:
          type: ["string", "null"]
    RunnerUpdate:
      type: object
      properties:
        name:
          type: string
        country:
          type: string
        countryCode:
          type: ["string", "null"]
        gender:
          type: string
        age:
          type: ["integer", "null"]
        bio:
          type: ["string", "null"]
    LeaderboardEntry:
      type: object
      required: [id, name, country, gender, rating, rank, totalRaces, ratingChange]
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
        countryCode:
          type: ["string", "null"]
        gender:
          type: string
        rating:
          type: number
        rank:
          type: integer
        ratingChange:
          type: number
        totalRaces:
          type: integer
        bestFinish:
          type: ["integer", "null"]
    Race:
      type: object
      required: [id, name, location, country, date, distanceKm, category, surface, status, difficultyScore]
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        country:
          type: string
        countryCode:
          type: ["string", "null"]
        date:
          type: string
        distanceKm:
          type: number
        distanceLabel:
          type: ["string", "null"]
        category:
          type: string
          enum: [50k, 100k, 50mi, 100mi, 200mi, other]
        surface:
          type: string
          enum: [trail, road, mountain, mixed]
        totalElevationM:
          type: ["integer", "null"]
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [upcoming, results_pending, completed]
        finishersCount:
          type: ["integer", "null"]
        weatherConditions:
          type: ["string", "null"]
          enum: [clear, rain, heat, snow, storm, null]
        technicalityRating:
          type: ["integer", "null"]
        difficultyScore:
          type: number
        organizerId:
          type: ["integer", "null"]
        organizerName:
          type: ["string", "null"]
        createdAt:
          type: string
...
    ResultWithRace:
      type: object
      required: [id, runnerId, raceId, points, race]
      properties:
        id:
          type: integer
        runnerId:
          type: integer
        raceId:
          type: integer
        position:
          type: ["integer", "null"]
        finishTimeSeconds:
          type: ["integer", "null"]
        dnf:
          type: boolean
        points:
          type: number
        race:
          $ref: "#/components/schemas/Race"
...
    RunnerStats:
      type: object
      required: [runnerId, totalRaces, totalFinishes, totalDnfs, totalDistanceKm, avgPosition, bestPosition, totalPoints]
      properties:
        runnerId:
          type: integer
        totalRaces:
          type: integer
        totalFinishes:
          type: integer
        totalDnfs:
          type: integer
        totalDistanceKm:
          type: number
        avgPosition:
          type: ["number", "null"]
        bestPosition:
          type: ["integer", "null"]
        totalPoints:
          type: number
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryBreakdown"
    CategoryBreakdown:
      type: object
      required: [category, races, avgPosition]
      properties:
        category:
          type: string
        races:
          type: integer
        avgPosition:
          type: ["number", "null"]
...
    ScrapeInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
    ScrapedResultItem:
      type: object
      required: [runnerName]
      properties:
        runnerName:
          type: string
        country:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        position:
          type: ["integer", "null"]
        finishTimeSeconds:
          type: ["integer", "null"]
        dnf:
          type: boolean
        birthYear:
          type: ["integer", "null"]
        ageCategory:
          type: ["string", "null"]
    ScrapePreview:
      type: object
      required: [source, url, totalFound, results]
      properties:
        raceName:
          type: ["string", "null"]
        raceDate:
          type: ["string", "null"]
        raceLocation:
          type: ["string", "null"]
        source:
          type: string
        url:
          type: string
        totalFound:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/ScrapedResultItem"
    ScrapeImportResponse:
      type: object
      required: [resultsCreated, runnersCreated, runnersUpdated, difficultyScore]
      properties:
        resultsCreated:
          type: integer
        runnersCreated:
          type: integer
        runnersUpdated:
          type: integer
        difficultyScore:
          type: number
        source:
          type: ["string", "null"]
        raceName:
          type: ["string", "null"]
    SearchCandidate:
      type: object
      required: [url, source, title, confidence]
      properties:
        url:
          type: string
        source:
          type: string
        title:
          type: string
        date:
          type: ["string", "null"]
        finishers:
          type: ["integer", "null"]
        confidence:
          type: string
          enum: [high, medium, low]
...
    AsyncJobStarted:
      type: object
      required: [jobId, total]
      properties:
        jobId:
          type: string
        total:
          type: integer
    ImportJob:
      type: object
      required: [id, status, processed, total]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [pending, running, done, error]
        processed:
          type: integer
        total:
          type: integer
        result:
          nullable: true
          $ref: "#/components/schemas/ScrapeImportResponse"
        error:
          type: ["string", "null"]
    AutoSearchResponse:
      type: object
      required: [candidates, raceName, raceDate]
      properties:
        raceName:
          type: string
        raceDate:
          type: string
        candidates:
          type: array
          items:
            $ref: "#/components/schemas/SearchCandidate"